+++ docs/md/releasenotes.md
@@ -1,3 +1,10 @@
+## Version 8.0.6
+
+### Improved
+
+- Add support for writing out zip64 archives. (`PGC-32`)
+- Improve speed for `ClassPool.contains` method. (`T5205`)
+
 ## Version 8.0.5
 
 ### Improved